backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
sentence-transformers>=2.2.0
networkx>=3.0
numpy>=1.24.0
scikit-learn>=1.3.0
spacy>=3.7.0
requests>=2.31.0
pydantic>=2.0.0
python-multipart>=0.0.6

